Fix corner case in TcpServer accept handler.
If the TcpServer accept handler gets called after the the server
has been shut down i.e. acceptor_ is NULL, it sets session state
to established and then allows the newly allocated session to get
destroyed when the scoped pointer is destroyed.  This causes the
assertion in ~TcpSession on !established to fail.

Fix is to check if acceptor is NULL before marking the session as
established.

Change-Id: Iffa2dd96769100a77162ed771d66ca975eecdfb5