Fix handling of bgp message with length field of 0
The bgp parser did flag an error on seeing a message with length field
set to 0, but TcpMessageReader::OnRead got trapped in an infinite loop.

Fixed by making TcpMessageReader::ReceiveCallback return success/failure
and by aborting processing in TcpMessageReader::OnRead when the callback
returns failure. Assumption is that the callback would trigger close and
eventually delete the session when it returns failure.

Add appropriate unit tests for bgp and tcp.

Change-Id: Iafc5c74899104e42aa05d561aab63611e657c747
Closes-Bug: #1408751