Fix issue with redirecting sandesh request in universal_parse.xsl
In http introspect, on clicking the hyperlink in the sandesh response
(field with link annotation), appropriate sandesh request is not sent to
the backend. The value of the link annotation field in the response
is not used to send the sandesh request. This was broken with the
recent commit to handle the special characters in the sandesh request.

Change-Id: I137de16ae7d4d9ba70ec33439afd0fcd1184fe02
Closes-Bug: #1418895