Second set of changes for bgp authentication via tcp-md5.
Take care of review comments regarding const and static. Some modularization.
Code to clear the listen socket when the peer is going away or peer address is
changing. Check for non-zero IP before setsockopt. Check if acceptor is present
before using it. Break up the code doing the setsockopt into one for the
session and one for the listen socket.

Change-Id: I72aa07e052c8905b5d6696040c4810bb8d8d5a4d
Partial-Bug: 1383393