Add SubnetParse routines that return (IP and MASK) from a string.
Today, the PrefixParse routines return the IP as it is. But, in some cases, we
need a routine that returns the 'bitwise and' of IP and Mask. Adding 2 such
SubnetParse routines.

1. Start using SubnetParse instead of PrefixParse in all the relevant locations.
2. Add new test file address_test.cc under src/net/test to test both, the
PrefixParse and SubnetParse, routines with both positive and negative tests.
3. Fix all the other tests to adhere to the use the new SubnetParse routines,
change tests to make them better and add new tests.
4. Add prefix length checks in Ip4PrefixParse().

Closes-Bug: #1512830

Change-Id: Iccf96f3c7df8ce500a3f2b406d01d170c0291bd0