1. Derived Stats Attributes
For DerivedStats, we will support attributes directly in the UVE Struct only.
These attributes can be be either elemental types, or maps with key and string,
and value as elemental type.

In the elemental case, we will instantiate a single DerivedStats object for the attribute

In the map case, we will instantiate a single DerivedStats object per entry in the map.
These DerivedStats objects will be identified by the same string keys as used in the raw map.

struct SandeshUVEData {
    1: string name (key="ObjectGeneratorInfo")
    2: optional bool deleted
    3: optional i32 x
    4: optional derived_stats_results.NullResult null_x (stats="x:")
    7: optional i32 y
    8: optional map<string,i32> tsm
    9: optional map<string,derived_stats_results.NullResult> null_tsm (stats="tsm:")
}

2. Periodic UVEs and attribute types

We has previously added the "period" annotation for UVE structs, for implementing
non-inline stats.
We now formalize the treatment of each optional attribute in the UVE struct as follows:
a) Inline. When client updates this attribute using UVE-Send, it is sent to the collector immidiately
b) Hidden . This attribute is shown in Introspect only. It is never sent to the collector.
   For this we use an annotation : (hidden="yes")
   This works for both "period"-annotated UVE struct, and those which are not
   It does not apply to DerivedStats
c) Periodic. Only valid for UVE structs with the period notation.
   These attributes are not sent inline when the client calls UVE-Send. Instead, they are sent
   periodically as per the "period" annotation of the UVE struct
   All DerivedStats (and only DerivedStats) in a "period"-annotated UVE struct will be periodic

struct SandeshPeriodicData {
    1: string name (key="ObjectGeneratorInfo")
    2: optional bool deleted
    3: optional i32 x (hidden="yes")
    4: optional derived_stats_results.SumResult null_x (stats="x:")
    5: optional map<string,i32> tsm
    6: optional map<string,derived_stats_results.SumResult> null_tsm (stats="tsm:")
} (period="60")

Change-Id: I289662f2d0279d4b489ea84c3128167b9fc68772
Partial-Bug: 1428271