Send elements in ContrailConfig as sorted OrderedDict
elements field in ContrailConfig (dynamicuve) is of type map. standard python
dictionary is unordered. It depends on the order in which the entries are
added/removed in the dictionary and the hash function. Therefore, if multiple
generators send the same map in different order, the UVE aggregation code
reports it as different value. Hence send the elements field in dynamicuve as
sorted OrderedDict.

Change-Id: Ice5d0f149e48e4b7eadd1fe2f6243fa72edc20c6
Closes-Bug: #1592755