C++ sandesh client library changes to accept more than 2 collectors
Presently, the contrail services may be either provisoned with the
collector list or it gets the collector list from discovery service. In
either case, the sandesh client library doesn't accept more than 2
collectors.
With the removal of discovery service, all the contrail services that
connects to collector would be provisioned with the collector list.
Therefore, it is necessary that the sandesh client library accepts
collector list without any limitation.
With this patch, InitGenerator() can accept collector list > 2 and there
is the notion of primary and secondary collector has been removed. The
sandesh client would connect to the first collector in the list and upon
connection failure/closure would connect to the next collector in the list
and so on. It is expected that the sandesh clients would randomize the
collector list before passing it to InitGenerator() and
ReConfigCollectors().
Renamed EvDiscUpdate event to EvCollectorUpdate. EvCollectorUpdate would
be enqueued upon receiving either discovery update or the
ReConfigCollectors() [called if the process receives SIGHUP to indicate
change in the collector list]
This patch doesn't remove the code that subscribes for the collector
service with discovery as this would break the functionality without the
provisioning/controller changes.

Change-Id: Ib84df4a3103e5483a9f3c7cb8ba5f3d034651058
Closes-Bug: #1641846