Do not free entry from values_ vector in index map until bit is freed
IndexMap stores values inside a vector values_ with the positions in the array
managed via a bitmap. IOW, a bit in the bitmap corresponds to an index/position
in the vector. Hence, the value inside the vector must be reset only when the
corresponding bit in the bitmap is reset.

With mvpn changes, bit reset was postponed to Resetbit() but the value was
reset from the vector inline in Remove() API itself. This can cause the
bit-map and vector to go out of sync if new insertions and deletions come in
between.

Verified using existing bgp_xmpp_deferq_test itself. More specific tests for
class IndexMap shall be added in a subsequent commit

Change-Id: I5b6c9d2ba52994d0ad9817480b0ae7f75a190824
Partial-Bug: 1734860