Revert removing staging of address.h
Basically a revert of change nr 39006

It was removed because this file had two Install() calls that
caused a minor warning. Unfortunetly the other call is in
contrail-analytics - a repository that is not downloaded on
Windows. Also contrail-common is a more basic dependency so
it should have this call and contrail-analytics should not.

Change-Id: I6b3b279c955d16250b5c3fae709fdc9e082a145e
Partial-Bug: #1737177