Change SCons dependency to file instead of directory
SCons does not support directories as targets, and they lead to hard to
track bugs. Instead of depending on the directory, depend on the tarball
being created there.

Also, add an ":sdist" target for pysandesh, to match other sdist
targets.

This is part of a patch series rewriting cfgm_common to tox:

  1. fix SCons directory dependencies <-- here
  2. add support for generic scons targets to rules.py
  3. rewrite cfgm_common to use tox and add generic targets
  4. remove rules.py support for using files as dependencies

Depends-On: Ic7fe446b13dc91587292237ed665be2fb2e09a0a
Change-Id: Ie459d7472e5f670ba94439de6173ffc1e9ea1f93
Partial-Bug: #1751438