Ignore Draft objects on the control node
Security policy drafts needs to be supported for auidits and validation before
applying them.
The security policy draft objects are not needed on the control node nor on the
compute nodes hence the best option is to ignore them on the control node.

o Added a flag in the ConfigCassandra parse context to indicate weather the
  parents FQ Name contains the pre-defined substring.
o After parsing all properties, if the the Draft flag is true, the object is
  deleted and FQName entry purged if present.
Depends-On: Iceb46ac2babb6f5d319f5b9bcc5c2050684de39c
Change-Id: I8292efbd7eba5d75c564c812e9234ef1ee54bfc0
Partial-Bug: #1755294