Gracefully ignore Draft objects on the control node
Security policy drafts needs to be supported for auidits and validation before
applying them.
The security policy draft objects are not needed on the control node nor on the
compute nodes hence the best option is to ignore them on the control node as
early as possible.
Purge FQNameCache and ObjectCache while ignoring to ensure that any updates
dont cause inadvertent configuration from being pushed to the backend.
Gracefully handle deletes and transformation of draft object to normal objects

Change-Id: Ib3e870623eda51e7ac7805c96cd678eafe767894
Partial-Bug: #1755294