Logging of sessions
This fix lets the session logs go to any or all of collector/
file/syslog. To accommodate this, a new logger object is
created for handling only SLO session destinations. The
destinations to this logger are provided by the agent. A snippet
of the generated_file is added to the bug.
Partial-Bug: 1729812

Change-Id: Ie6024b95aa5becb56291cd1ef229ad7915a0a57a