JSON encoding of the sandesh UVE messages
This fix adds the file wihch provides implementation of
transport API's exposed. The transport API's will be called
by the generated sandesh code.

UT added for the JSON generator. Currently the validity of
the JSON was checked with online tools. Holding the
verification part until we export a JSON library to sandesh.
Partial-Bug: 1789312

Change-Id: I1a8ddcdbccb5e8a899c7f45f3ba805eed65b5766