JSON encoding of the sandesh UVE messages
This fix adds the file wihch provides implementation of
transport API's exposed. The transport API's will be called
by the generated sandesh code.

UT added for the JSON generator. Currently the validity of
the JSON was checked with online tools. Holding the
verification part until we export a JSON library to sandesh.

Code Change:
JSON encoding unlike the XML encoding uses comma seperation
for members, string values for keys and {} enclosure for
complex types. To accomodate this following data structure
were introduced:

A vector to keep track of the element we are currently encoding
and whether its a first member of the familt or not:
Here is an example:
If we are encoding a list within sandesh structure,
current_sandesh_context_ maintains the following members in its
list viz., sandesh->struct_s1->list_l1
is_first_element_list_ maintains whether we are processing first
list element or we not.If we were in the middle of the list,
the structure is_first_element_list_ will be as follows:
true->false->false
Here the first false, indicates that list is not the first
member of the struct and the second false indicates that we
are done encoding the first element of the list.

Partial-Bug: 1789312

Change-Id: I1a8ddcdbccb5e8a899c7f45f3ba805eed65b5766