Remove sandesh log that prints change in sending level
This message is logged when the sandesh log level to the collector
changes due to sandesh queue reaching the high/low watermark. This
message is currently logged at level SYS_INFO. This message is not
useful and can flood the log file if the log_level is set to SYS_INFO or
above and hence removed.

Change-Id: I16d1b07f0174cca89029880114d09060824a9c53
Closes-Bug: #1790737