Control node not processing ETCD Watch Response with more than one event:
The issue was that we were not calling the callback fn for each event in the
watch response and instead overwriting the previous event. In older version of
contrail-go, etcd_notifer was used which resulted in only one event per
watch response. With the move to sync process, this changed and exposed the
issue. Fix is to invoke callback for each event in the watch response.

Change-Id: I6926970a41583ea348ce17a2f0293ee5d879dc8f
Partial-Bug: 1778805