Fix wrong usage of std::string::compare while creating URL for ifmap client
std::string::compare function behaves different if strings to be compared are not equal.
For this case we need to define position of element in original string to start comparison
and number of symbols to be compared(common prefix length for both strings).
This change fixes original issue when result is always not equal to 0.

Change-Id: Ia130c6a9578bb5d07719a5160644146b31dce589
Closes-bug: #1799922