Make task_annotations_test not flaky
There are two modifications:

TaskAnnotationsTest - access to count_ in Consume is not synchronized.
As two invocations of Consume may run in parallel and decrementation
is not atomic operation, result may be different than original author
expected.

Please remember that this modification fixes false negatives. Even
when policies created in SetUp are removed, this test is still likely
to pass.

TaskAnnotationsDeathTest - this test is executed only if
ConcurrencyChecker::enable_ is set to true (environment variable
CONCURRENCY_CHECK_ENABLE is set to "true"). Otherwise the test
can not pass.

Change-Id: If4b47ac1c4d579fe3d2fe101defbe8648b15c564
Closes-Jira-Bug: WIND-192