Contrail Query Engine memory leak.
Memory leak when query-engine tries to connect to non-operational cassandra in a loop.
Since the code in execution is cassandra cpp driver code, I have raised a bug in cpp-driver also. - https://datastax-oss.atlassian.net/browse/CPP-757
As of now, I've made some change in cql client code which brings down the leak significantly, but it is not zero.

DocImpact
closes-jira-bug: JCB-218701

Change-Id: Ie3c77f5a04b283b4131de5c9cff6f123fd88400c