Update patch set 3
Patch Set 3:

(14 comments)

I've put in two classes of comments:

1) nits, style ones (eg, const, methods probably not necessary? ...).

2) structural.

For the structural ones, I see two themes here:

a) I'd like to decouple sandesh and stats collection from the main objects. Both sandesh and stats collection should be _optional_ features, that for contrail customers get by default.
I have a proposal on how to achieve that in the big comment below. It boils down to having a simple FlagManager, with just Set and Get methods, and helper classes on the side that implement the "when controller send flags, call Set", "when flag is changed, send it to stats".

b) My original design was using MakeFlag as I was expecting to have flags templated on a type (eg, bool flags, string flags, ...) and using a C++11 pattern with move constructors. I suspect that if we follow along with 1, we don't need to use it here.

Patch-set: 3
Reviewer: Gerrit User 787 <787@c72cd27d-1743-4a55-84a4-cd13b8400c52>