RabbitMq client should delete the queue
Here are the steps to reproduce the issue:
1. stop the control node process (not kill it)
2. restart api server which should result in some config being pushed
3. restart control node

RabbitMq client tries to delete 'old' queue at restart but only
if it is not empty and not in use. In some cases, queue was not
empty because of which it does not get deleted. Earlier, with
exclusive queue flag, we were not able to use it further because
it was locked by previous connection. After changing the flag
to non-exclusive, DeclareQueue function gives an error if flags
are changed on the queue. This happens in the case where some config
gets pushed in the queue during restart of control node.
With this fix, code will "always" delete old queue even if there are
messages sitting in it. With this change, old queue will always
get deleted.

Change-Id: Id606c844e69807193b0fbf8620c6c20ec7b7d1d1
Closes-Jira-Bug: CC-323