Contrail collector is initializing state due to cassandra connection down.
In a multi-node setup, if first cassandra node is down then collector
doesn't create tables on second node after restart. It is happening
because there is a strict check to create tables only on first node.
This check is needed to handle only a very rare scenario. Under normal
circumstances, Cassandra nodes should be able to sync up with each other
on table creation. So, we are removing that single node schema creation code.

closes-jira-bug: CEM-10426
Change-Id: Ie14627f9061873ef60313898ad9ab4ca093a905e