Fix to return the new size of trace buffer when queried by applications.
The change in trace buffer capacity is correctly updated in the trace
container. But it is not udpated in a variable in TraceBuffer object.
Applications use this to query the size before request for the trace
contents. Fixed this issue by updating the information in TraceBuffer
object in addition to actually updating the Trace contrainer capacity.

Closes Jira Bug: CEM-11283

Change-Id: I75bfee966796f7e9d6764de342d14d186eda5bea