Add support for versioning in the c++ library.
Feature flag configuration can send multiple versions of the same
feature flag. This usually happens during an upgrade wheere different
modules come up at different points in time. In such a case, it is
possible that a module with an older version gets feature flag
configuration for a newer version and may not be ready to handle the
feature. To prevent the module from processing the new configuration, it
will compare its current version with the version field in the flag
configuration and only process those which match its version.

Closes-Jira-Bug: CEM-12360
Change-Id: I509143bba73adfeb6e280c84628c906fe39d8575